+++ priv/repo/migrations/20260921000700_add_last_heartbeat_at_to_access_keys.exs
@@ -1,0 +1,9 @@
+defmodule Apiary.Repo.Migrations.AddLastHeartbeatAtToAccessKeys do
+  use Ecto.Migration
+
+  def change do
+    alter table(:access_keys) do
+      add :last_heartbeat_at, :utc_datetime_usec
+    end
+  end
+end